+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 0.11.2 (unreleased)
 
+  * allow JPI/HPI file extension and packaging type to be configured by the `fileExtension` property, which will default
+    to `hpi`
+    ([JENKINS-28408](https://issues.jenkins-ci.org/browse/JENKINS-28408))
+
 ## 0.11.1 (2015-05-08)
 
   * changed packaging in generated POM back to jpi to fix a regression bug